scene drawer added
