+++ public/app/features/dashboard-scene/SceneDrawer.tsx
@@ -1,27 +1,23 @@
-import { css } from '@emotion/css';
 import React from 'react';
 
-import { GrafanaTheme2 } from '@grafana/data';
 import { SceneComponentProps, SceneObjectBase, SceneObject, SceneObjectState } from '@grafana/scenes';
-import { Drawer, useStyles2 } from '@grafana/ui';
-import appEvents from 'app/core/app_events';
-import { ShowModalReactEvent } from 'app/types/events';
+import { Drawer } from '@grafana/ui';
+import { Props as DrawerProps } from '@grafana/ui/src/components/Drawer/Drawer';
+
+import appEvents from '../../core/app_events';
+import { ShowModalReactEvent } from '../../types/events';
 
 export type SceneDrawerProps = {
   scene: SceneObject;
-  title: string;
-  onDismiss: () => void;
-};
+  onClose: () => void;
+} & Partial<Omit<DrawerProps, 'onClose'>>;
 
 export function SceneDrawer(props: SceneDrawerProps) {
-  const { scene, title, onDismiss } = props;
-  const styles = useStyles2(getStyles);
+  const { scene, title, onClose, size = 'lg', ...rest } = props;
 
   return (
-    <Drawer title={title} onClose={onDismiss} size="lg">
-      <div className={styles.drawerInnerWrapper}>
-        <scene.Component model={scene} />
-      </div>
+    <Drawer title={title} onClose={onClose} {...rest} size={size}>
+      <scene.Component model={scene} />
     </Drawer>
   );
 }
@@ -47,18 +43,4 @@ export function launchSceneDrawerInGlobalModal(props: Omit<SceneDrawerProps, 'on
   };
 
   appEvents.publish(new ShowModalReactEvent(payload));
-}
-
-function getStyles(theme: GrafanaTheme2) {
-  return {
-    drawerInnerWrapper: css({
-      display: 'flex',
-      padding: theme.spacing(2),
-      background: theme.isDark ? theme.colors.background.canvas : theme.colors.background.primary,
-      position: 'absolute',
-      left: 0,
-      right: 0,
-      top: 0,
-    }),
-  };
 }
