+++ public/app/features/trails/Integrations/DataTrailEmbedded.tsx
@@ -1,7 +1,9 @@
+import { css } from '@emotion/css';
 import React from 'react';
 
-import { AdHocVariableFilter } from '@grafana/data';
+import { AdHocVariableFilter, GrafanaTheme2 } from '@grafana/data';
 import { SceneComponentProps, SceneObjectBase, SceneObjectState, SceneTimeRangeLike } from '@grafana/scenes';
+import { useStyles2 } from '@grafana/ui';
 
 import { DataTrail } from '../DataTrail';
 
@@ -23,7 +25,13 @@ export class DataTrailEmbedded extends SceneObjectBase<DataTrailEmbeddedState> {
 }
 
 function DataTrailEmbeddedRenderer({ model }: SceneComponentProps<DataTrailEmbedded>) {
-  return <model.trail.Component model={model.trail} />;
+  const styles = useStyles2(getStyles);
+
+  return (
+    <div className={styles.drawerInnerWrapper}>
+      <model.trail.Component model={model.trail} />
+    </div>
+  );
 }
 
 export function buildDataTrailFromState({ metric, filters, dataSourceUid, timeRange }: DataTrailEmbeddedState) {
@@ -34,4 +42,18 @@ export function buildDataTrailFromState({ metric, filters, dataSourceUid, timeRa
     initialFilters: filters,
     embedded: true,
   });
+}
+
+function getStyles(theme: GrafanaTheme2) {
+  return {
+    drawerInnerWrapper: css({
+      display: 'flex',
+      padding: theme.spacing(2),
+      background: theme.isDark ? theme.colors.background.canvas : theme.colors.background.primary,
+      position: 'absolute',
+      left: 0,
+      right: 0,
+      top: 0,
+    }),
+  };
 }
